add test folder for H100
