+++ .buildkite/pipeline.yml
@@ -38,7 +38,7 @@ steps:
     commands:
       - export VLLM_LOGGING_LEVEL=DEBUG
       - export VLLM_WORKER_MULTIPROC_METHOD=spawn
-      - pytest -s -v tests/multi_stages/test_qwen2_5_omni.py
+      - pytest -s -v tests/multi_stages/
     agents:
       queue: "gpu_4_queue" # g6.4xlarge instance on AWS, has 1 L4 GPU
     plugins:
@@ -51,12 +51,12 @@ steps:
           volumes:
             - "/fsx/hf_cache:/fsx/hf_cache"
 
-  - label: "Qwen3-Omni Model Test"
+  - label: "Omni Model Test with H100"
     timeout_in_minutes: 20
     depends_on: image-build
     commands:
       - export VLLM_WORKER_MULTIPROC_METHOD=spawn
-      - pytest -s -v tests/multi_stages/test_qwen3_omni.py
+      - pytest -s -v tests/multi_stages_h100/
     agents:
       queue: "mithril-h100-pool"
     plugins:
