--- tests/multi_stages/test_qwen3_omni.py
@@ -1,159 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-"""
-E2E tests for Qwen3-Omni model with audio/video/image input and audio output.
-"""
-
-import os
-from pathlib import Path
-
-import pytest
-from vllm.assets.audio import AudioAsset
-from vllm.assets.image import ImageAsset
-from vllm.assets.video import VideoAsset
-from vllm.multimodal.image import convert_image_mode
-
-from .conftest import OmniRunner
-
-os.environ["VLLM_WORKER_MULTIPROC_METHOD"] = "spawn"
-
-models = ["Qwen/Qwen3-Omni-30B-A3B-Instruct"]
-
-# CI stage config for 2xH100-80G GPUs
-CI_STAGE_CONFIG_PATH = str(Path(__file__).parent / "stage_configs" / "qwen3_omni_ci.yaml")
-
-
-@pytest.mark.parametrize("model", models)
-def test_audio_to_audio(omni_runner: type[OmniRunner], model: str) -> None:
-    """Test processing audio, generating audio output."""
-    with omni_runner(model, seed=42, stage_configs_path=CI_STAGE_CONFIG_PATH) as runner:
-        # Prepare inputs
-        question = "What is recited in the audio?"
-        audio = AudioAsset("mary_had_lamb").audio_and_sample_rate
-        audio = (audio[0][: 16000 * 5], audio[1])  # Trim to first 5 seconds
-
-        outputs = runner.generate_multimodal(
-            prompts=question,
-            audios=audio,
-        )
-
-        # Verify we got outputs from multiple stages
-        assert len(outputs) > 0
-
-        # Find and verify text output (thinker stage)
-        text_output = None
-        for stage_output in outputs:
-            if stage_output.final_output_type == "text":
-                text_output = stage_output
-                break
-
-        assert text_output is not None
-        assert len(text_output.request_output) > 0
-        text_content = text_output.request_output[0].outputs[0].text
-        assert text_content is not None
-        assert len(text_content.strip()) > 0
-
-        # Find and verify audio output (code2wav stage)
-        audio_output = None
-        for stage_output in outputs:
-            if stage_output.final_output_type == "audio":
-                audio_output = stage_output
-                break
-
-        assert audio_output is not None
-        assert len(audio_output.request_output) > 0
-
-        # Verify audio tensor exists and has content
-        audio_tensor = audio_output.request_output[0].multimodal_output["audio"]
-        assert audio_tensor is not None
-        assert audio_tensor.numel() > 0
-
-
-@pytest.mark.parametrize("model", models)
-def test_video_to_audio(omni_runner: type[OmniRunner], model: str) -> None:
-    """Test processing video, generating audio output."""
-    with omni_runner(model, seed=42, stage_configs_path=CI_STAGE_CONFIG_PATH) as runner:
-        # Prepare inputs
-        question = "Describe the video briefly."
-        video = VideoAsset(name="baby_reading", num_frames=4).np_ndarrays
-
-        outputs = runner.generate_multimodal(
-            prompts=question,
-            videos=video,
-        )
-
-        # Verify we got outputs from multiple stages
-        assert len(outputs) > 0
-
-        # Find and verify text output (thinker stage)
-        text_output = None
-        for stage_output in outputs:
-            if stage_output.final_output_type == "text":
-                text_output = stage_output
-                break
-
-        assert text_output is not None
-        assert len(text_output.request_output) > 0
-        text_content = text_output.request_output[0].outputs[0].text
-        assert text_content is not None
-        assert len(text_content.strip()) > 0
-
-        # Find and verify audio output (code2wav stage)
-        audio_output = None
-        for stage_output in outputs:
-            if stage_output.final_output_type == "audio":
-                audio_output = stage_output
-                break
-
-        assert audio_output is not None
-        assert len(audio_output.request_output) > 0
-
-        # Verify audio tensor exists and has content
-        audio_tensor = audio_output.request_output[0].multimodal_output["audio"]
-        assert audio_tensor is not None
-        assert audio_tensor.numel() > 0
-
-
-@pytest.mark.parametrize("model", models)
-def test_image_to_audio(omni_runner: type[OmniRunner], model: str) -> None:
-    """Test processing video, generating audio output."""
-    with omni_runner(model, seed=42, stage_configs_path=CI_STAGE_CONFIG_PATH) as runner:
-        # Prepare inputs
-        question = "What is in this image?"
-        image = convert_image_mode(ImageAsset("cherry_blossom").pil_image.resize((128, 128)), "RGB")
-
-        outputs = runner.generate_multimodal(
-            prompts=question,
-            images=image,
-        )
-
-        # Verify we got outputs from multiple stages
-        assert len(outputs) > 0
-
-        # Find and verify text output (thinker stage)
-        text_output = None
-        for stage_output in outputs:
-            if stage_output.final_output_type == "text":
-                text_output = stage_output
-                break
-
-        assert text_output is not None
-        assert len(text_output.request_output) > 0
-        text_content = text_output.request_output[0].outputs[0].text
-        assert text_content is not None
-        assert len(text_content.strip()) > 0
-
-        # Find and verify audio output (code2wav stage)
-        audio_output = None
-        for stage_output in outputs:
-            if stage_output.final_output_type == "audio":
-                audio_output = stage_output
-                break
-
-        assert audio_output is not None
-        assert len(audio_output.request_output) > 0
-
-        # Verify audio tensor exists and has content
-        audio_tensor = audio_output.request_output[0].multimodal_output["audio"]
-        assert audio_tensor is not None
-        assert audio_tensor.numel() > 0
